+++ src/main/java/org/logstash/filters/geoip/GeoIPFilter.java
@@ -20,12 +20,15 @@ package org.logstash.filters.geoip;
 
 import com.maxmind.db.CHMCache;
 import com.maxmind.db.InvalidDatabaseException;
+import com.maxmind.db.Network;
 import com.maxmind.geoip2.exception.AddressNotFoundException;
 import com.maxmind.geoip2.exception.GeoIp2Exception;
+import com.maxmind.geoip2.model.AnonymousIpResponse;
 import com.maxmind.geoip2.model.AsnResponse;
 import com.maxmind.geoip2.model.CityResponse;
 import com.maxmind.geoip2.model.CountryResponse;
 import com.maxmind.geoip2.model.DomainResponse;
+import com.maxmind.geoip2.model.EnterpriseResponse;
 import com.maxmind.geoip2.model.IspResponse;
 import com.maxmind.geoip2.record.*;
 import org.apache.logging.log4j.LogManager;
@@ -35,37 +38,23 @@ import org.logstash.Event;
 import com.maxmind.geoip2.DatabaseReader;
 import org.logstash.ext.JrubyEventExtLibrary.RubyEvent;
 
+import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.*;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 
-public class GeoIPFilter {
-  private static Logger logger = LogManager.getLogger();
-  // The free GeoIP2 databases
-  private static final String CITY_LITE_DB_TYPE = "GeoLite2-City";
-  private static final String COUNTRY_LITE_DB_TYPE = "GeoLite2-Country";
-  private static final String ASN_LITE_DB_TYPE = "GeoLite2-ASN";
-
-  // The paid GeoIP2 databases
-  private static final String CITY_DB_TYPE = "GeoIP2-City";
-  private static final String CITY_AFRICA_DB_TYPE = "GeoIP2-City-Africa";
-  private static final String CITY_ASIA_PACIFIC_DB_TYPE = "GeoIP2-City-Asia-Pacific";
-  private static final String CITY_EUROPE_DB_TYPE = "GeoIP2-City-Europe";
-  private static final String CITY_NORTH_AMERICA_DB_TYPE = "GeoIP2-City-North-America";
-  private static final String CITY_SOUTH_AMERICA_DB_TYPE = "GeoIP2-City-South-America";
-  private static final String COUNTRY_DB_TYPE = "GeoIP2-Country";
-  private static final String ISP_DB_TYPE = "GeoIP2-ISP";
-  private static final String DOMAIN_DB_TYPE = "GeoIP2-Domain";
-
+public class GeoIPFilter implements Closeable {
+  private static final Logger logger = LogManager.getLogger();
   private final String sourceField;
   private final String targetField;
-  private final Set<Fields> desiredFields;
+  private final Set<Field> desiredFields;
+  private final Database database;
   private final DatabaseReader databaseReader;
-
-  private final Function<Fields,String> fieldReferenceExtractor;
+  private final Function<Field,String> fieldReferenceExtractor;
 
   public GeoIPFilter(String sourceField, String targetField, List<String> fields, String databasePath, int cacheSize,
                      String ecsCompatibility) {
@@ -73,31 +62,43 @@ public class GeoIPFilter {
     this.targetField = targetField;
     switch (ecsCompatibility) {
       case "disabled":
-        this.fieldReferenceExtractor = Fields::getFieldReferenceLegacy;
+        this.fieldReferenceExtractor = Field::getFieldReferenceLegacy;
         break;
       case "v1":
       case "v8":
-        this.fieldReferenceExtractor = Fields::getFieldReferenceECSv1;
+        this.fieldReferenceExtractor = Field::getFieldReferenceECSv1;
         break;
       default:
         throw new UnsupportedOperationException("Unknown ECS version " + ecsCompatibility);
     }
 
-    final File database = new File(databasePath);
+    final File databaseFile = new File(databasePath);
     try {
-      this.databaseReader = new DatabaseReader.Builder(database).withCache(new CHMCache(cacheSize)).build();
+      this.databaseReader = new DatabaseReader.Builder(databaseFile).withCache(new CHMCache(cacheSize)).build();
     } catch (InvalidDatabaseException e) {
       throw new IllegalArgumentException("The database provided is invalid or corrupted.", e);
     } catch (IOException e) {
       throw new IllegalArgumentException("The database provided was not found in the path", e);
     }
+
+    this.database = getDatabase(databaseReader);
     this.desiredFields = createDesiredFields(fields, !ecsCompatibility.equals("disabled"));
+  }
+
+  private static Database getDatabase(DatabaseReader reader) {
+    final String databaseType = reader.getMetadata().getDatabaseType();
+    final Database database = Database.fromDatabaseType(databaseType);
+
+    if (database == Database.UNKNOWN) {
+      logger.warn("The provided database type {} is not supported", databaseType);
+    }
+
+    return database;
   }
 
   public static boolean isDatabaseValid(String databasePath) {
     final File database = new File(databasePath);
-    try {
-      new DatabaseReader.Builder(database).build();
+    try (DatabaseReader ignore = new DatabaseReader.Builder(database).build()) {
       return true;
     } catch (InvalidDatabaseException e) {
       logger.debug("The database provided is invalid or corrupted");
@@ -107,38 +108,31 @@ public class GeoIPFilter {
     return false;
   }
 
-  private Set<Fields> createDesiredFields(List<String> fields, final boolean ecsCompatibilityEnabled) {
-    Set<Fields> desiredFields = EnumSet.noneOf(Fields.class);
-    if (fields == null || fields.isEmpty()) {
-      switch (databaseReader.getMetadata().getDatabaseType()) {
-        case CITY_LITE_DB_TYPE:
-        case CITY_DB_TYPE:
-        case CITY_AFRICA_DB_TYPE:
-        case CITY_ASIA_PACIFIC_DB_TYPE:
-        case CITY_EUROPE_DB_TYPE:
-        case CITY_NORTH_AMERICA_DB_TYPE:
-        case CITY_SOUTH_AMERICA_DB_TYPE:
-          desiredFields = ecsCompatibilityEnabled ? Fields.DEFAULT_ECS_CITY_FIELDS : Fields.DEFAULT_CITY_FIELDS;
-          break;
-        case COUNTRY_LITE_DB_TYPE:
-        case COUNTRY_DB_TYPE:
-          desiredFields = Fields.DEFAULT_COUNTRY_FIELDS;
-          break;
-        case ISP_DB_TYPE:
-          desiredFields = Fields.DEFAULT_ISP_FIELDS;
-          break;
-        case ASN_LITE_DB_TYPE:
-          desiredFields = Fields.DEFAULT_ASN_LITE_FIELDS;
-          break;
-        case DOMAIN_DB_TYPE:
-          desiredFields = Fields.DEFAULT_DOMAIN_FIELDS;
-      }
-    } else {
-      for (String fieldName : fields) {
-        desiredFields.add(Fields.parseField(fieldName));
-      }
+  private Set<Field> createDesiredFields(List<String> fields, final boolean ecsCompatibilityEnabled) {
+    if (fields != null && !fields.isEmpty()) {
+      return fields.stream()
+              .map(Field::parseField)
+              .collect(Collectors.toCollection(() -> EnumSet.noneOf(Field.class)));
     }
-    return desiredFields;
+
+    if (database == Database.CITY) {
+      return createCityDefaultFields(ecsCompatibilityEnabled);
+    }
+
+    return database.getDefaultFields();
+  }
+
+  private Set<Field> createCityDefaultFields(boolean ecsCompatibilityEnabled) {
+    // When ECS is disabled, change the default region code field from REGION_ISO_CODE to
+    // REGION_CODE (BC)
+    if (!ecsCompatibilityEnabled) {
+      final EnumSet<Field> ecsDisabledFields = EnumSet.copyOf(database.getDefaultFields());
+      ecsDisabledFields.remove(Field.REGION_ISO_CODE);
+      ecsDisabledFields.add(Field.REGION_CODE);
+      return ecsDisabledFields;
+    }
+
+    return database.getDefaultFields();
   }
 
   public boolean handleEvent(RubyEvent rubyEvent) {
@@ -162,32 +156,31 @@ public class GeoIPFilter {
       return false;
     }
 
-    Map<Fields, Object> geoData = new HashMap<>();
+    Map<Field, Object> geoData = new HashMap<>();
 
     try {
       final InetAddress ipAddress = InetAddress.getByName(ip);
-      switch (databaseReader.getMetadata().getDatabaseType()) {
-        case CITY_LITE_DB_TYPE:
-        case CITY_DB_TYPE:
-        case CITY_AFRICA_DB_TYPE:
-        case CITY_ASIA_PACIFIC_DB_TYPE:
-        case CITY_EUROPE_DB_TYPE:
-        case CITY_NORTH_AMERICA_DB_TYPE:
-        case CITY_SOUTH_AMERICA_DB_TYPE:
+      switch (database) {
+        case CITY:
           geoData = retrieveCityGeoData(ipAddress);
           break;
-        case COUNTRY_LITE_DB_TYPE:
-        case COUNTRY_DB_TYPE:
+        case COUNTRY:
           geoData = retrieveCountryGeoData(ipAddress);
           break;
-        case ASN_LITE_DB_TYPE:
+        case ASN:
           geoData = retrieveAsnGeoData(ipAddress);
           break;
-        case ISP_DB_TYPE:
+        case ISP:
           geoData = retrieveIspGeoData(ipAddress);
           break;
-        case DOMAIN_DB_TYPE:
+        case DOMAIN:
           geoData = retrieveDomainGeoData(ipAddress);
+          break;
+        case ENTERPRISE:
+          geoData = retrieveEnterpriseGeoData(ipAddress);
+          break;
+        case ANONYMOUS_IP:
+          geoData = retrieveAnonymousIpGeoData(ipAddress);
           break;
         default:
           throw new IllegalStateException("Unsupported database type " + databaseReader.getMetadata().getDatabaseType() + "");
@@ -203,7 +196,7 @@ public class GeoIPFilter {
     return applyGeoData(geoData, event);
   }
 
-  private boolean applyGeoData(Map<Fields, Object> geoData, Event event) {
+  private boolean applyGeoData(Map<Field, Object> geoData, Event event) {
     if (geoData == null) {
       return false;
     }
@@ -217,8 +210,8 @@ public class GeoIPFilter {
     }
 
     String targetFieldReference = "[" + this.targetField + "]";
-    for (Map.Entry<Fields, Object> it: geoData.entrySet()) {
-      final Fields field = it.getKey();
+    for (Map.Entry<Field, Object> it: geoData.entrySet()) {
+      final Field field = it.getKey();
       final String subFieldReference = this.fieldReferenceExtractor.apply(field);
 
       if (subFieldReference.equals("[]")) {
@@ -230,7 +223,7 @@ public class GeoIPFilter {
     return true;
   }
 
-  private Map<Fields,Object> retrieveCityGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
+  private Map<Field,Object> retrieveCityGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
     CityResponse response = databaseReader.city(ipAddress);
     Country country = response.getCountry();
     City city = response.getCity();
@@ -238,7 +231,7 @@ public class GeoIPFilter {
     Continent continent = response.getContinent();
     Postal postal = response.getPostal();
     Subdivision subdivision = response.getMostSpecificSubdivision();
-    Map<Fields, Object> geoData = new HashMap<>();
+    Map<Field, Object> geoData = new EnumMap<>(Field.class);
 
     // if location is empty, there is no point populating geo data
     // and most likely all other fields are empty as well
@@ -246,104 +239,95 @@ public class GeoIPFilter {
       return geoData;
     }
 
-    for (Fields desiredField : this.desiredFields) {
+    for (Field desiredField : this.desiredFields) {
       switch (desiredField) {
         case CITY_NAME:
           String cityName = city.getName();
           if (cityName != null) {
-            geoData.put(Fields.CITY_NAME, cityName);
+            geoData.put(Field.CITY_NAME, cityName);
           }
           break;
         case CONTINENT_CODE:
           String continentCode = continent.getCode();
           if (continentCode != null) {
-            geoData.put(Fields.CONTINENT_CODE, continentCode);
+            geoData.put(Field.CONTINENT_CODE, continentCode);
           }
           break;
         case CONTINENT_NAME:
           String continentName = continent.getName();
           if (continentName != null) {
-            geoData.put(Fields.CONTINENT_NAME, continentName);
+            geoData.put(Field.CONTINENT_NAME, continentName);
           }
           break;
         case COUNTRY_NAME:
           String countryName = country.getName();
           if (countryName != null) {
-            geoData.put(Fields.COUNTRY_NAME, countryName);
+            geoData.put(Field.COUNTRY_NAME, countryName);
           }
           break;
         case COUNTRY_CODE2:
           String countryCode2 = country.getIsoCode();
           if (countryCode2 != null) {
-            geoData.put(Fields.COUNTRY_CODE2, countryCode2);
+            geoData.put(Field.COUNTRY_CODE2, countryCode2);
           }
           break;
         case COUNTRY_CODE3:
           String countryCode3 = country.getIsoCode();
           if (countryCode3 != null) {
-            geoData.put(Fields.COUNTRY_CODE3, countryCode3);
+            geoData.put(Field.COUNTRY_CODE3, countryCode3);
           }
           break;
         case IP:
-          geoData.put(Fields.IP, ipAddress.getHostAddress());
+          geoData.put(Field.IP, ipAddress.getHostAddress());
           break;
         case POSTAL_CODE:
           String postalCode = postal.getCode();
           if (postalCode != null) {
-            geoData.put(Fields.POSTAL_CODE, postalCode);
+            geoData.put(Field.POSTAL_CODE, postalCode);
           }
           break;
         case DMA_CODE:
           Integer dmaCode = location.getMetroCode();
           if (dmaCode != null) {
-            geoData.put(Fields.DMA_CODE, dmaCode);
+            geoData.put(Field.DMA_CODE, dmaCode);
           }
           break;
         case REGION_NAME:
           String subdivisionName = subdivision.getName();
           if (subdivisionName != null) {
-            geoData.put(Fields.REGION_NAME, subdivisionName);
+            geoData.put(Field.REGION_NAME, subdivisionName);
           }
           break;
         case REGION_CODE:
           String subdivisionCode = subdivision.getIsoCode();
           if (subdivisionCode != null) {
-            geoData.put(Fields.REGION_CODE, subdivisionCode);
+            geoData.put(Field.REGION_CODE, subdivisionCode);
           }
           break;
         case REGION_ISO_CODE:
-          String countryCodeForRegion = country.getIsoCode();
-          String regionCode2 = subdivision.getIsoCode();
-          if (countryCodeForRegion != null && regionCode2 != null) {
-            geoData.put(Fields.REGION_ISO_CODE, String.format("%s-%s", countryCodeForRegion, regionCode2));
-          }
+          parseRegionIsoCodeField(country, subdivision)
+                  .ifPresent(data -> geoData.put(Field.REGION_ISO_CODE, data));
           break;
         case TIMEZONE:
           String locationTimeZone = location.getTimeZone();
           if (locationTimeZone != null) {
-            geoData.put(Fields.TIMEZONE, locationTimeZone);
+            geoData.put(Field.TIMEZONE, locationTimeZone);
           }
           break;
         case LOCATION:
-          Double latitude = location.getLatitude();
-          Double longitude = location.getLongitude();
-          if (latitude != null && longitude != null) {
-            Map<String, Object> locationObject = new HashMap<>();
-            locationObject.put("lat", latitude);
-            locationObject.put("lon", longitude);
-            geoData.put(Fields.LOCATION, locationObject);
-          }
+          parseLocationField(location)
+                  .ifPresent(data -> geoData.put(Field.LOCATION, data));
           break;
         case LATITUDE:
           Double lat = location.getLatitude();
           if (lat != null) {
-            geoData.put(Fields.LATITUDE, lat);
+            geoData.put(Field.LATITUDE, lat);
           }
           break;
         case LONGITUDE:
           Double lon = location.getLongitude();
           if (lon != null) {
-            geoData.put(Fields.LONGITUDE, lon);
+            geoData.put(Field.LONGITUDE, lon);
           }
           break;
       }
@@ -352,33 +336,33 @@ public class GeoIPFilter {
     return geoData;
   }
 
-  private Map<Fields,Object> retrieveCountryGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
+  private Map<Field,Object> retrieveCountryGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
     CountryResponse response = databaseReader.country(ipAddress);
     Country country = response.getCountry();
     Continent continent = response.getContinent();
-    Map<Fields, Object> geoData = new HashMap<>();
+    Map<Field, Object> geoData = new EnumMap<>(Field.class);
 
-    for (Fields desiredField : this.desiredFields) {
+    for (Field desiredField : this.desiredFields) {
       switch (desiredField) {
         case IP:
-          geoData.put(Fields.IP, ipAddress.getHostAddress());
+          geoData.put(Field.IP, ipAddress.getHostAddress());
           break;
         case COUNTRY_CODE2:
           String countryCode2 = country.getIsoCode();
           if (countryCode2 != null) {
-            geoData.put(Fields.COUNTRY_CODE2, countryCode2);
+            geoData.put(Field.COUNTRY_CODE2, countryCode2);
           }
           break;
         case COUNTRY_NAME:
           String countryName = country.getName();
           if (countryName != null) {
-            geoData.put(Fields.COUNTRY_NAME, countryName);
+            geoData.put(Field.COUNTRY_NAME, countryName);
           }
           break;
         case CONTINENT_NAME:
           String continentName = continent.getName();
           if (continentName != null) {
-            geoData.put(Fields.CONTINENT_NAME, continentName);
+            geoData.put(Field.CONTINENT_NAME, continentName);
           }
           break;
       }
@@ -387,37 +371,37 @@ public class GeoIPFilter {
     return geoData;
   }
 
-  private Map<Fields, Object> retrieveIspGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
+  private Map<Field, Object> retrieveIspGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
     IspResponse response = databaseReader.isp(ipAddress);
 
-    Map<Fields, Object> geoData = new HashMap<>();
-    for (Fields desiredField : this.desiredFields) {
+    Map<Field, Object> geoData = new EnumMap<>(Field.class);
+    for (Field desiredField : this.desiredFields) {
       switch (desiredField) {
         case IP:
-          geoData.put(Fields.IP, ipAddress.getHostAddress());
+          geoData.put(Field.IP, ipAddress.getHostAddress());
           break;
         case AUTONOMOUS_SYSTEM_NUMBER:
           Integer asn = response.getAutonomousSystemNumber();
           if (asn != null) {
-            geoData.put(Fields.AUTONOMOUS_SYSTEM_NUMBER, asn);
+            geoData.put(desiredField, asn);
           }
           break;
         case AUTONOMOUS_SYSTEM_ORGANIZATION:
           String aso = response.getAutonomousSystemOrganization();
           if (aso != null) {
-            geoData.put(Fields.AUTONOMOUS_SYSTEM_ORGANIZATION, aso);
+            geoData.put(desiredField, aso);
           }
           break;
         case ISP:
           String isp = response.getIsp();
           if (isp != null) {
-            geoData.put(Fields.ISP, isp);
+            geoData.put(Field.ISP, isp);
           }
           break;
         case ORGANIZATION:
           String org = response.getOrganization();
           if (org != null) {
-            geoData.put(Fields.ORGANIZATION, org);
+            geoData.put(Field.ORGANIZATION, org);
           }
           break;
       }
@@ -426,24 +410,31 @@ public class GeoIPFilter {
     return geoData;
   }
 
-  private Map<Fields, Object> retrieveAsnGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
+  private Map<Field, Object> retrieveAsnGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
     AsnResponse response = databaseReader.asn(ipAddress);
-    Map<Fields, Object> geoData = new HashMap<>();
-    for (Fields desiredField : this.desiredFields) {
+    Network network = response.getNetwork();
+
+    Map<Field, Object> geoData = new EnumMap<>(Field.class);
+    for (Field desiredField : this.desiredFields) {
       switch (desiredField) {
         case IP:
-          geoData.put(Fields.IP, ipAddress.getHostAddress());
+          geoData.put(Field.IP, ipAddress.getHostAddress());
           break;
         case AUTONOMOUS_SYSTEM_NUMBER:
           Integer asn = response.getAutonomousSystemNumber();
           if (asn != null) {
-            geoData.put(Fields.AUTONOMOUS_SYSTEM_NUMBER, asn);
+            geoData.put(Field.AUTONOMOUS_SYSTEM_NUMBER, asn);
           }
           break;
         case AUTONOMOUS_SYSTEM_ORGANIZATION:
           String aso = response.getAutonomousSystemOrganization();
           if (aso != null) {
-            geoData.put(Fields.AUTONOMOUS_SYSTEM_ORGANIZATION, aso);
+            geoData.put(Field.AUTONOMOUS_SYSTEM_ORGANIZATION, aso);
+          }
+          break;
+        case NETWORK:
+          if (network != null) {
+            geoData.put(Field.NETWORK, network.toString());
           }
           break;
       }
@@ -452,18 +443,199 @@ public class GeoIPFilter {
     return geoData;
   }
 
-  private Map<Fields, Object> retrieveDomainGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
+  private Map<Field, Object> retrieveDomainGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
     DomainResponse response = databaseReader.domain(ipAddress);
-    Map<Fields, Object> geoData = new HashMap<>();
-    for (Fields desiredField : this.desiredFields) {
+    Map<Field, Object> geoData = new EnumMap<>(Field.class);
+    for (Field desiredField : this.desiredFields) {
       switch (desiredField) {
         case DOMAIN:
           String domain = response.getDomain();
-          geoData.put(Fields.DOMAIN, domain);
+          geoData.put(Field.DOMAIN, domain);
           break;
       }
     }
 
     return geoData;
+  }
+
+  private Map<Field, Object> retrieveEnterpriseGeoData(InetAddress ipAddress) throws GeoIp2Exception, IOException {
+    EnterpriseResponse response = databaseReader.enterprise(ipAddress);
+
+    Map<Field, Object> geoData = new EnumMap<>(Field.class);
+    Country country = response.getCountry();
+    City city = response.getCity();
+    Location location = response.getLocation();
+    Continent continent = response.getContinent();
+    Subdivision subdivision = response.getMostSpecificSubdivision();
+
+    Integer asn = response.getTraits().getAutonomousSystemNumber();
+    String organizationName = response.getTraits().getAutonomousSystemOrganization();
+    Network network = response.getTraits().getNetwork();
+
+    boolean isHostingProvider = response.getTraits().isHostingProvider();
+    boolean isTorExitNode = response.getTraits().isTorExitNode();
+    boolean isAnonymousVpn = response.getTraits().isAnonymousVpn();
+    boolean isAnonymous = response.getTraits().isAnonymous();
+    boolean isPublicProxy = response.getTraits().isPublicProxy();
+    boolean isResidentialProxy = response.getTraits().isResidentialProxy();
+
+    for (Field desiredField : this.desiredFields) {
+      switch (desiredField) {
+        case IP:
+          geoData.put(Field.IP, ipAddress.getHostAddress());
+          break;
+        case COUNTRY_CODE2:
+          String countryIsoCode = country.getIsoCode();
+          if (countryIsoCode != null) {
+            geoData.put(desiredField, countryIsoCode);
+          }
+          break;
+        case COUNTRY_NAME:
+          String countryName = country.getName();
+          if (countryName != null) {
+            geoData.put(desiredField, countryName);
+          }
+          break;
+        case CONTINENT_NAME:
+          String continentName = continent.getName();
+          if (continentName != null) {
+            geoData.put(desiredField, continentName);
+          }
+          break;
+        case REGION_ISO_CODE:
+          parseRegionIsoCodeField(country, subdivision)
+                  .ifPresent(data -> geoData.put(desiredField, data));
+          break;
+        case REGION_NAME:
+          String subdivisionName = subdivision.getName();
+          if (subdivisionName != null) {
+            geoData.put(desiredField, subdivisionName);
+          }
+          break;
+        case CITY_NAME:
+          String cityName = city.getName();
+          if (cityName != null) {
+            geoData.put(desiredField, cityName);
+          }
+          break;
+        case TIMEZONE:
+          String locationTimeZone = location.getTimeZone();
+          if (locationTimeZone != null) {
+            geoData.put(desiredField, locationTimeZone);
+          }
+          break;
+        case LOCATION:
+          parseLocationField(location)
+                  .ifPresent(data -> geoData.put(desiredField, data));
+          break;
+        case AUTONOMOUS_SYSTEM_NUMBER:
+          if (asn != null) {
+            geoData.put(desiredField, asn);
+          }
+          break;
+        case AUTONOMOUS_SYSTEM_ORGANIZATION:
+          if (organizationName != null) {
+            geoData.put(desiredField, organizationName);
+          }
+          break;
+        case NETWORK:
+          if (network != null) {
+            geoData.put(desiredField, network.toString());
+          }
+          break;
+        case HOSTING_PROVIDER:
+          geoData.put(desiredField, isHostingProvider);
+          break;
+        case TOR_EXIT_NODE:
+          geoData.put(desiredField, isTorExitNode);
+          break;
+        case ANONYMOUS_VPN:
+          geoData.put(desiredField, isAnonymousVpn);
+          break;
+        case ANONYMOUS:
+          geoData.put(desiredField, isAnonymous);
+          break;
+        case PUBLIC_PROXY:
+          geoData.put(desiredField, isPublicProxy);
+          break;
+        case RESIDENTIAL_PROXY:
+          geoData.put(desiredField, isResidentialProxy);
+          break;
+      }
+    }
+    return geoData;
+  }
+
+  private Map<Field, Object> retrieveAnonymousIpGeoData(final InetAddress ipAddress) throws GeoIp2Exception, IOException {
+    AnonymousIpResponse response = databaseReader.anonymousIp(ipAddress);
+
+    Map<Field, Object> geoData = new EnumMap<>(Field.class);
+    boolean isHostingProvider = response.isHostingProvider();
+    boolean isTorExitNode = response.isTorExitNode();
+    boolean isAnonymousVpn = response.isAnonymousVpn();
+    boolean isAnonymous = response.isAnonymous();
+    boolean isPublicProxy = response.isPublicProxy();
+    boolean isResidentialProxy = response.isResidentialProxy();
+
+    for (Field desiredField : this.desiredFields) {
+      switch (desiredField) {
+        case IP:
+          geoData.put(desiredField, ipAddress.getHostAddress());
+          break;
+        case HOSTING_PROVIDER:
+          geoData.put(desiredField, isHostingProvider);
+          break;
+        case TOR_EXIT_NODE:
+          geoData.put(desiredField, isTorExitNode);
+          break;
+        case ANONYMOUS_VPN:
+          geoData.put(desiredField, isAnonymousVpn);
+          break;
+        case ANONYMOUS:
+          geoData.put(desiredField, isAnonymous);
+          break;
+        case PUBLIC_PROXY:
+          geoData.put(desiredField, isPublicProxy);
+          break;
+        case RESIDENTIAL_PROXY:
+          geoData.put(desiredField, isResidentialProxy);
+          break;
+      }
+    }
+    return geoData;
+  }
+
+  private Optional<Map<String, Object>> parseLocationField(Location location) {
+    Double latitude = location.getLatitude();
+    Double longitude = location.getLongitude();
+    if (latitude != null && longitude != null) {
+      Map<String, Object> locationObject = new HashMap<>();
+      locationObject.put("lat", latitude);
+      locationObject.put("lon", longitude);
+      return Optional.of(locationObject);
+    }
+
+    return Optional.empty();
+  }
+
+  private Optional<String> parseRegionIsoCodeField(final Country country, final Subdivision subdivision) {
+    String countryCodeForRegion = country.getIsoCode();
+    String regionCode2 = subdivision.getIsoCode();
+    if (countryCodeForRegion != null && regionCode2 != null) {
+      return Optional.of(String.format("%s-%s", countryCodeForRegion, regionCode2));
+    }
+
+    return Optional.empty();
+  }
+
+  @Override
+  public void close() {
+    if (databaseReader != null) {
+      try {
+        databaseReader.close();
+      } catch (IOException e) {
+        // Ignore
+      }
+    }
   }
 }